Update 3.8 Performing Subqueries
The previous task saw you write a complex query with multiple joins and filters. Now your manager wants you to analyze the results of that query. The only catch is that revising your query could take quite some time, not to mention the risk of breaking it. Instead, you decide to use it as a subquery (or inner query) to answer the business questions listed in the task.